==> notebooks/01_data_overview.ipynb <==
# 01 data overview

The goal of this project is to use classification models to predict student graduation status and to identify the key factors influencing graduation. By evaluating different classification models and analyzing feature importance, the project aims to provide actionable recommendations to help institutions improve graduation rates and better support at-risk students.


import pandas as pd


df1 = pd.read_csv("data/raw/Cohort_3.csv")

df2 = pd.read_csv("data/raw/Cohort_3_DA.csv")

df = pd.concat([df1, df2])
df.tail()

1. <b>Schema Checkes; data types

df.info()

#drop the id no columns
df.drop(columns=['Id. No', 'ID No.'], inplace=True)

df.head(2)

#Covert Timestamp to datetime from a string
import datetime as dt
df["Timestamp"] = pd.to_datetime(df["Timestamp"])
df.dtypes

df.head(2)

2. <b>Renaming the columns

df.columns

new_columns = [
    "timestamp",
    "age_range",
    "gender",
    "country",
    "referral_source",
    "years_experience",
    "track_applied",
    "weekly_commitment_hours",
    "main_aim",
    "motivation",
    "skill_level",
    "aptitude_test_completed",
    "total_score",
    "graduated"
]
df.columns = new_columns

df.head()

3. <b>Categorical Standardization

#Keeep the years_experience ordered and categorical for visualization
experience_categories = [
    "Less than six months",
    "6 months - 1 year",
    "1-3 years",
    "4-6 years"
]

df["years_experience"] = pd.Categorical(
    df["years_experience"], 
    categories=experience_categories, 
    ordered=True
)

#ordered categories
commitment_categories = ["less than 6 hours", "7-14 hours", "more than 14 hours"]

df["weekly_commitment_hours"] = pd.Categorical(
    df["weekly_commitment_hours"], 
    categories=commitment_categories, 
    ordered=True
)

We have overlaps like "Upskill" and "both upskilling and connecting with fellow data professionals".
To make analysis meaningful, I could consolidate:

Upskill → includes "Upskill" and "both upskilling and connecting..."

Learn afresh → "Learn data afresh"

Portfolio → "Build a project portfolio"

Networking → "Connect with fellow data professionals"


def clean_main_aim(x):
    if "upskill" in x.lower():
        return "Upskill"
    elif "afresh" in x.lower():
        return "Learn afresh"
    elif "portfolio" in x.lower():
        return "Portfolio"
    elif "connect" in x.lower():
        return "Networking"
    else:
        return "Other"

df["main_aim"] = df["main_aim"].apply(clean_main_aim)
df["main_aim"].value_counts()

The raw labels are long and slightly overlapping and unordered. A cleaned version might look like:

Beginner → "Beginner - I have NO learning or work experience"

Elementary → "Elementary - I have theoretical understanding of basics"

Intermediate → "Intermediate - I have theoretical knowledge + experience"

skill_level_categories = ["Beginner", "Elementary", "Intermediate"]

df["skill_level"] = pd.Categorical(
    df["skill_level"].replace({
        "Beginner - I have NO learning or work experience in data analysis/ data science": "Beginner",
        "Elementary - I have theoretical understanding of basic data analysis/ data science concepts": "Elementary",
        "Intermediate - I have theoretical knowledge and experience in data analysis/ data science": "Intermediate"
    }),
    categories=skill_level_categories,
    ordered=True
)
df["skill_level"].value_counts()

We can kee age ranges are ordinal categories, so the goal is to represent them in a way that shows the distribution across age groups clearly.

age_categories = ["18-24 years", "25-34 years", "35-44 years", "45-54 years"]

df["age_range"] = pd.Categorical(df["age_range"], categories=age_categories, ordered=True)

df.head()

df.info()

Save the cleaned data ready for Exploratory data analysis

df.to_csv("data/interim/cleaned_df.csv", index=False)
